[Docs] 과제 ex06~ex07 수정

+++ ex06/move_dir.sh
@@ -1,1 +1,7 @@
-mv test.txt ../newdir/
+FILE=test.txt
+if [ -e $FILE ]; 
+then
+	mv test.txt ../newdir/
+else
+	echo "파일이 없는데"
+fi
